+++ package.json
@@ -25,6 +25,11 @@
       "import": "./build/client.mjs",
       "types": "./build/client.d.ts"
     },
+    "./primitives": {
+      "require": "./build/primitives.js",
+      "import": "./build/primitives.mjs",
+      "types": "./build/primitives.d.ts"
+    },
     "./index.css": "./build/ui.css"
   },
   "scripts": {
