+++ src/components/primitives/index.ts
@@ -1,0 +1,23 @@
+export { Portal } from "@ark-ui/react";
+
+export {
+  default as Drawer,
+  DrawerBackdrop,
+  DrawerTrigger,
+  DrawerContainer,
+  DrawerContent,
+  DrawerCloseTrigger,
+  DrawerTitle,
+  DrawerDescription,
+} from "./Drawer/Drawer";
+
+export type {
+  DrawerProps,
+  DrawerTriggerProps,
+  DrawerBackdropProps,
+  DrawerContainerProps,
+  DrawerContentProps,
+  DrawerCloseTriggerProps,
+  DrawerTitleProps,
+  DrawerDescriptionProps,
+} from "./Drawer/Drawer";
