+++ src/index.primitives.ts
@@ -1,0 +1,3 @@
+"use client";
+
+export * from "components/primitives";
