+++ tsup.config.ts
@@ -5,7 +5,11 @@ import { defineConfig as defineTsupConfig } from "tsup";
  * @see https://tsup.egoist.dev
  */
 const tsupConfig = defineTsupConfig({
-  entry: { ui: "src/index.ts", client: "src/index.client.ts" },
+  entry: {
+    ui: "src/index.ts",
+    client: "src/index.client.ts",
+    primitives: "src/index.primitives.ts",
+  },
   // TODO enable Rollup tree-shaking, temporarily disabled to allow properly injecting `use client` directive banner (Rollup tree-shaking removes it); track https://github.com/egoist/tsup/issues/835
   // treeshake: true,
   sourcemap: true,
